Make non-root user non-optional
If someone really wants to run as root they should know enough to modify
the instructions to do so: so make the creation of a non-root user
non-optional.

Also some of the later commands already assume that everything is
running under a non-root user, so requiring it makes more sense later.